9-21 school side
Noticed and started to work thru the issues with some listed lat and lon being incorrect or not plotting correctly. Want to ensure that this works, as making this iterative means when I add other lat and lon to the excel it will still run itself. Need to begin considering if I wish to use a different method for getting the data I want out of the system (assuming I can find it...)