9-27 Home side
Completed functionality of making a list of GEO locations, started building a framework for making a csv file to import and use as a list to pull from the HPC files. Sent via email a txt outlining pulling files from the HPC, not clear how to pull iterativelly if possible.