10-5 school side
Successful in formatting the HPC document, gotten out the time series for each location. Should potentially refine so that the method works for all .nc file geographies. For Master plot, added Wilmont index (NEED TO DOUBLE CHECK). Found a range of parameters on the cold side that allow the appearance of a seemingly gaussian distribution for all parameters. It appears to improve the R-squared, but at what cost? Need to look at the other metrics as well to see what parts improve or get worse. Need to add the new fitting metrics to a table for assessment of all PFT's. Need to apply this new allowed range to all PFT's to see differences. Begin analysis of time series.