10-18 Home side
Created the template doc for the new 3way ANOVA. Performed the previous method on PSIIGEO to compare sample instance to grand instance (See ANOVA doc on GDrive) Need to formalize the classification levels for the other conditions. I think I will just make a few and if any of them suck ignore them.